Update __init__.py
Parameter size change with v10 firmware

+++ zhaquirks/inovelli/__init__.py
@@ -88,7 +88,7 @@ class Inovelli_VZM31SN_Cluster(CustomCluster):
             0x000E: ("default_level_remote", t.uint8_t, True),
             0x000F: ("state_after_power_restored", t.uint8_t, True),
             0x0011: ("load_level_indicator_timeout", t.uint8_t, True),
-            0x0012: ("active_power_reports", t.uint16_t, True),
+            0x0012: ("active_power_reports", t.uint8_t, True),
             0x0013: ("periodic_power_and_energy_reports", t.uint8_t, True),
             0x0014: ("active_energy_reports", t.uint16_t, True),
             0x0015: ("power_type", t.uint8_t, True),
